auxiliary function to get the length of a string

+++ 0x12-singly_linked_lists/_strlen.c
@@ -5,10 +5,10 @@
   * @str: the string
   * Return: the length of the string
   */
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
-	int i;
-	int len = 0;
+	size_t i;
+	size_t len = 0;
 
 	for (i =  0; str[i] != '\0'; i++)
 	{
